Gubanov_dz # 3.1- float update

+++ dz3.1_gubanov.py
@@ -2,16 +2,16 @@ number1 = input("Enter a number '1': ")
 number2 = input("Enter a number '2': ")
 user_action = input("Enter an action (*,/,+,-): ")
 if user_action == "*":
-    result = float(int( number1 )*int(number2))
+    result = float(float( number1 )*float(number2))
 elif user_action == "+":
-    result = float(int( number1 )+int(number2))
+    result = float(float( number1 )+float(number2))
 elif user_action == "-":
-    result = float(int(number1) - int(number2))
+    result = float(float(number1) - float(number2))
 elif user_action == "/":
     if number2  == "0":
         result = "Zero division prohibited.\nPut a correct data!"
     else:
-        result = float(int(number1) / int(number2))
+        result = float(float(number1) / float(number2))
 elif user_action != "*,/,+,-":
     result = "Types prohibited.\nPut a correct data for action!"
